@@ -1,6 +1,8 @@
 class HomeController < ApplicationController
   def index
   	@message = Message.new
+    @previous_meetups = Archive.published.limit(2).all
+    @upcoming_meetup = Archive.unpublished.first
   	check_for_signups
   end
 
@@ -11,7 +13,7 @@ private
   	if params[:message].present?
 
 	    @message = Message.new(params[:message])
-	    
+
 	    if @message.valid?
 	    	mimi = MadMimi.new(ENV["MIMI_USERNAME"], ENV["MIMI_API"])
       	mimi.add_to_list(@message.email,'MotionMeetup')
@@ -20,7 +22,7 @@ private
 	      flash[:error] = "Please fill all fields correctly."
 	    end
 
-	end
+	  end
   end
 
 end
