@@ -2,18 +2,32 @@
 #
 # Table name: users
 #
-#  id         :integer          not null, primary key
-#  name       :string(255)
-#  email      :string(255)
-#  provider   :string(255)
-#  uid        :string(255)
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
+#  id                     :integer          not null, primary key
+#  name                   :string(255)      default("")
+#  email                  :string(255)      default(""), not null
+#  provider               :string(255)
+#  uid                    :string(255)
+#  created_at             :datetime
+#  updated_at             :datetime
+#  encrypted_password     :string(255)      default(""), not null
+#  reset_password_token   :string(255)
+#  reset_password_sent_at :datetime
+#  remember_created_at    :datetime
+#  sign_in_count          :integer          default(0)
+#  current_sign_in_at     :datetime
+#  last_sign_in_at        :datetime
+#  current_sign_in_ip     :string(255)
+#  last_sign_in_ip        :string(255)
 #
 
 class User < ActiveRecord::Base
-  attr_accessible :provider, :uid, :name, :email
-  validates_presence_of :name
+  # No new user signups allowed :P
+  if Rails.env.production?
+    devise :database_authenticatable, :recoverable, :rememberable, :trackable, :validatable
+  else
+    devise :database_authenticatable, :recoverable, :rememberable, :trackable, :validatable, :registerable
+  end
+
 
   def self.create_with_omniauth(auth)
     create! do |user|
