@@ -6,8 +6,6 @@ Motionmeetup::Application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
@@ -32,23 +30,37 @@ Motionmeetup::Application.configure do
   }
 
 
-
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
 
   # Only use best-standards-support built into browsers
   config.action_dispatch.best_standards_support = :builtin
 
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
-
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
 
   # Do not compress assets
   config.assets.compress = false
 
   # Expands the lines which load the assets
   config.assets.debug = true
+
+  config.eager_load = false
+
+  # create a connection
+  # connection = Fog::Storage.new({
+  #   :provider                 => 'AWS',
+  #   :aws_access_key_id        => YOUR_AWS_ACCESS_KEY_ID,
+  #   :aws_secret_access_key    => YOUR_AWS_SECRET_ACCESS_KEY
+  # })
+
+  #config.paperclip_defaults = {:storage => :fog, :fog_credentials => {:provider => "Local", :local_root => "#{Rails.root}/public"}, :fog_directory => "", :fog_host => "localhost"}
+
+  # config.paperclip_defaults = {
+  #   :storage => :filesystem,
+  #   :url  => "/uploads/:attachment/:id/:style/:basename-:fingerprint.:extension",
+  #   :path => Rails.root + "public/uploads/:attachment/:id/:style/:basename-:fingerprint.:extension" }
+
+
+
+  #:endpoint => Rails.root.join 'tmp'
+
 end
