@@ -11,9 +11,6 @@ Motionmeetup::Application.configure do
   config.serve_static_assets = true
   config.static_cache_control = "public, max-age=3600"
 
-  # Log error messages when you accidentally call methods on nil
-  config.whiny_nils = true
-
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
@@ -29,13 +26,12 @@ Motionmeetup::Application.configure do
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
 
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
 
   # Print deprecation notices to the stderr
   config.active_support.deprecation = :stderr
 
   # ActionMailer Config
   config.action_mailer.default_url_options = { :host => 'example.com' }
-
+  config.eager_load = false
+  PAPERCLIP_STORAGE_OPTIONS = {:storage => :filesystem}
 end
