@@ -1,8 +1,8 @@
 class CreateUsers < ActiveRecord::Migration
   def change
     create_table :users do |t|
-      t.string :name
-      t.string :email
+      t.string :name,  :null => true, :default => ""
+      t.string :email, :null => false, :default => ""
       t.string :provider
       t.string :uid
 
